pyproject.toml
[project]
name = "usas-validator"
version = "0.1.0"
description = "Validates USAS Tags"
authors = [
    { name = "UCREL Research Centre", email = "ucrel@lancaster.ac.uk" },
]
license = "Apache-2.0"
license-files = [
    "LICENSE",
]
keywords = [
    "usas",
    "USAS",
    "UCREL",
    "ucrel",
    "semantics",
    "evaluation"
]
classifiers = [
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]
requires-python = ">=3.10, <3.15"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0, <7.0",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.11.0"]
build-backend = "uv_build"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[dependency-groups]
dev = [
    "coverage[toml]>=7.13.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
    "ty>=0.0.13",
]


[tool.ruff.lint]
# Enable the isort rules.
extend-select = ["I"]

#[tool.ruff.lint.isort]
#known-third-party = ["src", "tests"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [
    "tests"
]

[tool.coverage.run]
branch = true
source = [
    "src"
]
command_line = "-m pytest"

[tool.coverage.report]
show_missing = true
